requirements.txt
python-nmap
colored
yaspin